added load balancing
